backend/weather/requirements.txt
asgiref==3.8.1
debugpy==1.8.5
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0
django-cors-headers==3.5.0
pika==1.1.0
psycopg2==2.9.9

#django-mysql==3.9
#mysqlclient==2.0.1
